move nemo file back into nemo folder
